pyproject.toml
[project]
name = "virt-graph"
version = "0.8.0"
description = "Graph-like queries over relational data using LLM reasoning"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9",
    "networkx>=3.2",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "faker>=24.0",
    "scipy (>=1.16.3,<2.0.0)",
    "neo4j (>=6.0.3,<7.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
neo4j = [
    "neo4j>=5.0",
]

[tool.poetry]
packages = [{include = "virt_graph", from = "src"}]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "mkdocs (>=1.6.1,<2.0.0)",
    "mkdocs-material (>=9.7.0,<10.0.0)"
]
